[ECS] Detect process hostname when deploying as a sidecar in ECS Managed Instances (#42461)
Fixes hostname detection for the process-component when running on ECS Managed Instances as a sidecar. The agent will set the host to `sidecar_host:TASK_ARN` and set the container host type to `ContainerHostType_sidecar`

When deploying as a sidecar in ECS, we don't want process/containers to report with the host tag. Instead, we want the host to be the task. This behavior is consistent with how we handle fargate hosts for process/containers.

- Deployed the datadog agent using image `datadog/agent-dev:stzou-exp-74-full` into ECS using Managed Instance as launch type with `DD_ECS_DEPLOYMENT_MODE=sidecar`.
- validated that live process/container data is reported with the proper tagging and hostname.
<img width="345" height="151" alt="image" src="https://github.com/user-attachments/assets/a68f4160-685b-4325-b6b3-eeec69046583" />

Co-authored-by: steve.zou <steve.zou@datadoghq.com>
(cherry picked from commit f196cbd322f2b6b62c10a44be7af31d86cf140f2)