add openssl3 generated files for darwin arm64 (#42731)
This change brings in the files captured during a configure/make of openssl3 for darwin arm64.
It is a prequel to https://github.com/DataDog/datadog-agent/pull/42016.
That PR will use the files in a meaningful way. This change is too large to examine in any useful way, so and #42016 can't work without these files, so I want to add these first for a rubber stamp review, so the next one can be a sensible review.

Co-authored-by: tony.aiuto <tony.aiuto@datadoghq.com>