write bool when false (#42604)
### What does this PR do?
ensure options explicitly set to `false` are written to the config

### Motivation
https://datadoghq.atlassian.net/browse/WINA-2008
yaml marshaller does not write false bools tagged with `omitempty`, must use `*bool` instead.

`remote_updates` defaults to `true` as of https://github.com/DataDog/datadog-agent/pull/42500, so we need a way to disable it

### Additional Notes
removed explicit `RemoteUpdates` value set in `setConfigInstallerDaemon`, it is now only written to config when explicitly provided. Updating the default to `false` for FIPS installs will have to be handled another way.

Co-authored-by: branden.clark <branden.clark@datadoghq.com>