dyninst/procsubscribe: fix symdb update handling bug (#42744)
### What does this PR do?

The logic in the state management was confusing the lack of a file with the lack of a matched path. This commit fixes the issue and adds testing.

### Motivation

We observed via the QA of the latest release that some paginated symdb uploads weren't completing. They were getting canceled after 5s (the poll interval for rc updates). 

### Describe how you validated your changes

Wrote tests.

### Additional Notes

For backport.

Co-authored-by: andrew.werner <andrew.werner@datadoghq.com>