fix: Revert the ddtrace hang workaround since it's resolved upstream (#42739)
### What does this PR do?

Reverts DataDog/datadog-agent#42222


This PR disables a the workaround for ddtrace hangs. The workaround is no longer necessary because we upgraded to [ddtrace 3.16.4](https://github.com/DataDog/dd-trace-py/releases/tag/v3.16.4) in integrations-core, which includes the proper fix.

### Motivation

Re-enable the features disabled in the workaround (Python wall clock and cpu times when profiling integrations).

### Describe how you validated your changes

We upgraded to ddtrace 3.16.4 in our CI, and disabled the workaround there. We haven't observed hangs.

### Additional Notes


Co-authored-by: enrico.donnici <enrico.donnici@datadoghq.com>