Check if Windows services unexpectedly restart in startstop tests (#42851)
### What does this PR do?
This PR checks  in E2E startstop test if a Windows service that was expected to stop unexpectedly restarted.  
This can happen if something causes service to auto-restart and during that transition the stop request was issued. The stop request is ignored, leasing to the service to go back to running.

### Motivation
https://datadoghq.atlassian.net/browse/WINA-1871

### Describe how you validated your changes
CI


Co-authored-by: alexander.nip <alexander.nip@datadoghq.com>