[AGENTCFG-405] Fetching instance-type from ec2 instances and adding it to the inventoryhost payload (#42701)
### What does this PR do?

Refer to [this
doc](https://docs.google.com/document/d/1lQiJII2b4R8TzvPElg9Q8ynMqdMg3GHapQ_rVQFmFAQ/edit?usp=sharing).
After talking to stakeholders, we have chosen to implement **option 3**;
the work for this option on the agent side is to fetch the instance type
from various cloud providers and report them in a new field in the
inventoryhost payload. This PR creates the scaffolding for the new
field, and does fetching for ec2 instances.

### Motivation

### Describe how you validated your changes

I ran an agent on ec2 with these changes incorporated, and when I run
`agent diagnose show-metadata inventory-host`, I see `instance-type` as
a field in the payload.

### Additional Notes