[ AGNTLOG-384 ] Shift the logs agent file tailer to a dependency-injected file opener (#42741)
### What does this PR do?
This PR focuses on shifting the file tailer (and by extension the file fingerprinter) over to a mockable file opener entity.
Additional code tweaks include: 
- The two WildcardModes in the file tailer have been added as constants
- Several test suites have modified test component constructors to better facilitate the new dependency injection

There is only one expected change in actual execution logic, which is the move from the fingerprinter to call OpenLogFile instead of OpenShared. This would qualify as a bug fix were the fingerprinting feature released. 

### Motivation
The primary purpose for this change is to better support modularity/mockability in the logs agent

### Describe how you validated your changes

### Additional Notes


Co-authored-by: ryan.hall <ryan.hall@datadoghq.com>