discovery: add ContainerID to WLM entities when process collection is disabled (#42426)
### What does this PR do?

This PR makes it so the container ID of a Process is added to its WLM entity when only Service collection is enabled.

### Motivation

Container ID is part of the expected Service Discovery payload. However, it is currently only added when regular process collection is enabled. This PR aims to fix this discrepancy.

### Describe how you validated your changes

Added unit tests cases for it.

### Additional Notes


Co-authored-by: vincent.whitchurch <vincent.whitchurch@datadoghq.com>