dyninst: move temporary files from /var/tmp to /tmp (#42602)
In our k8s Helm charts we mount some dirs under
`/var/tmp/datadog-agent/system-probe/...`, but we mount them as host dirs.
That's not what we want for the
`/var/tmp/datadog-agent/system-probe/dynamic-instrumentation` (we want
this dir to be an emptyDir mount so that it does not survive pod
restarts). So, let's move our caches out of `/var/tmp`, to make the
distinction more apparent.

Also replace underscores with dashes in some of the paths, for
consistency with other files.

Co-authored-by: andrei.matei <andrei.matei@datadoghq.com>