[omnibus] Rewrite sysctl-probe-offline-skip patch to OpenSCAP
This change rewrites the sysctl-probe-offline-skip patch
to OpenSCAP, so it returns SYSCHAR_FLAG_NOT_APPLICABLE
instead of OVAL_RESULT_TRUE when OpenSCAP is running
in offline mode.

Sysctl rules usually check for both the value present
in sysctl configuration files and sysctl value at runtime.
When running in offline mode, we chose to consider
only the value of the sysctl configuration files.