[WP] Add user session ID as part of the process_context_t in kernel events (#41124)
### What does this PR do?

This PR adds the user session ID in `process_context_t`. This changes means that we no longer need to wait for a process to exec (or fork) in order to start attaching the user session context to our events: as soon as it is available, it will automatically be attached to enrich our events.

### Motivation

Improve our user session tracking capabilities.

Co-authored-by: theop-dd <theo.putegnat@datadoghq.com>