dyninst/irgen: limit support explicitly to snapshots
Before this change, we'd happily take any probes that had
a where even if they weren't snapshot probes. This adds explicit
chacking to make sure we only support snapshot probes, and adds
facilities to test that throughout out testing infra. To do that
it leverages tags to mark probes that are going to have issues.