dyninst: fix bugs related to stack tracking for events (#42984)
### What does this PR do?

This PR fixes issues we've seen where stack traces are not being properly recorded. It also adds better observability into complex buffering scenarios, and fixes a bug when the ring buffer is full.

### Motivation

We've seen in the wild that stack traces are missing in some events where returns are also missing. This can be explained by situations where the entry event was buffered while a subsequent event corresponding to that entry stack was emitted. 

### Describe how you validated your changes

There's some testing.

### Additional Notes

Relates to https://datadoghq.atlassian.net/browse/DEBUG-4031

Co-authored-by: andrew.werner <andrew.werner@datadoghq.com>