Remove jemalloc (#43032)
### What does this PR do?

Effectively rolls back #29628.

### Motivation

It is proving difficult to increase the default compression level for metrics to anything higher than 1 as it can use more CPU to compress than is desired. It looks like the default level will remain at 1 for the foreseeable future. This means we won't be needing to change the default allocator, so `jemalloc` is just taking up space.

### Describe how you validated your changes

### Additional Notes


Co-authored-by: stephen.wakely <stephen.wakely@datadoghq.com>