[AMCC-149]Add FilterList to checks (#43024)
### What does this PR do?

Adds the filterList to filter metrics coming in from checks

### Motivation

This expands the filterList so we can filter custom metrics coming in from checks. For example via the `openmetrics` check.

### Describe how you validated your changes

I have tested manually as well as added a couple of unit tests.

### Additional Notes

Currently the functionality to receive the filterList from remote config is handled by the dogstatsd server. This made sense when the filter list was only operating on dogstatsd metrics. This functionality needs to be extracted into a shared component that both dogstatsd and the check sampler can access. A follow up PR will address this.


Co-authored-by: stephen.wakely <stephen.wakely@datadoghq.com>