[SBOM] add bom ref simplification pass (+ config option) (#43051)
### What does this PR do?

The bom refs in the SBOM payload are used as a kind of index, in order to maintain the tree relations between components (like dependency relations).

Bom refs are currently a simple copy of the purl (or a randomly generated UUID if no purl). As such they can use a great amount of memory in the total payload size. This PR adds a configurable pass mapping the bom refs to simple increasing integers, allowing to reduce greatly the amount of payload used by bom refs.

This pass is configurable through `sbom.simplify_bom_refs` and defaults to false for now, awaiting further testing.

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: paul.cacheux <paul.cacheux@datadoghq.com>