utils: fix tag name for pool telemetry (#43019)
### What does this PR do?

Changes the "name" tag for the pool telemetry to "pool_name", as "name" was conflicting with other infra tags.

### Motivation

Avoid conflicts

### Describe how you validated your changes

Locally tested that the tag name is changed.

### Additional Notes

No teams other that GPU monitoring (us) is currently using the pool telemetry so there's no one else to ping about this breaking change in the name.

Co-authored-by: guillermo.julian <guillermo.julian@datadoghq.com>