[ AGNTLOG-400 ] Provide sane default values for byte_checksum fingerprint strategy (#43089)
### What does this PR do?
The current fingerprint configuration allows a user to specify a "bytes_checksum" fingerprint strategy with no associated count configuration. This defaults the count value to 1, which is an extremely unreasonable value for a bytes_checksum to have. As a fix, we will now generate an appropriate default count value for all valid fingerprint strategies.

Note that the per-source fingerprint configuration continues to expect all configuration options to be defined in order to be valid.

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: ryan.hall <ryan.hall@datadoghq.com>