Add new CRD type in WLM, add api extentions informer in dca
In Workload-Meta add a new kind named `CRD` which holds the group,kind
and version of a given CRD found the cluster.

Add necessary method like dumping the string representation of that CRD
entity.

Add a new api-extensions informer that will allow us (later) to
list/inform on all CRDs on the cluster.

This will allows us to add a new collector that will be able to collect
all known CRDs on the cluster and push them to Workload-Meta for other
components.

Signed-off-by: Alexandre Lavigne <alexandre.lavigne@datadoghq.com>