chore(ssi): refactor autoinstrumentation tests (#43005)
### What does this PR do?
This commit refactors the auto instrumentation webhook tests so that they test the webhook in isolation without requiring access to private module resources.

### Motivation
We want to refactor this webhook as it has grown organically over several years (see [SSI Kubernetes | Platform Stability](https://docs.google.com/document/d/1NqrPEUn3RfcdS_hQUQB-pJFJN9N-x5I202CB7s7CnwQ/edit?usp=sharing)). As a first step, we want to ensure that the tests are solid at a high level so that refactors to the implementation are safer.

The goal was to provide config and a pod as input and then validate that the webhook modified the pod in ways we expect.

### Describe how you validated your changes
I went through every test case line by line and ensured we had identical coverage in the new test.

### Additional Notes
I want to move `integration_test.go` into `auto_instrumentation_test.go` but I wanted to make this PR readable. I will do that in a follow up PR.


Co-authored-by: mark.spicer <mark.spicer@datadoghq.com>