[CSPM] extract sysprobe client into interface and add local version (#43138)
### What does this PR do?

The CSPM agent can query the system-probe for some compliance data benchmarks. This PR extracts this capability into an interface, and adds a local version of this feature. This will allow the system-probe version of the CSPM agent to query the data locally instead of querying it through the endpoint.

No functional change intended.

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: paul.cacheux <paul.cacheux@datadoghq.com>