Use Linux image to execute E2E tests (#42832)
### What does this PR do?

Use classic Linux image for E2E tests. This will make the maintenance easier by no longer need to maintain a dedicated e2e tests image.
As a consequences we need to fetch the Pulumi plugins in the CI. It can be done easily in the go_e2e_deps jobs and the Pulumi are then cached.

### Motivation

Ease maintenance of e2e tests image. This is also an unblocker for merging test-infra-definitions inside datadog-agent

### Describe how you validated your changes

Complete CI running.

### Additional Notes


Co-authored-by: kevin.fairise <kevin.fairise@datadoghq.com>