CONS-7845 fix orchestrator check memory leak (#43123)
### What does this PR do?

This PR does
* `orchestrator.go` : Moved collector initialization after the leader check to avoid starting informer on non-leader pods.

* `terminated_resource_bundle.go` : Replaced [:0] with make() to ensure consistent  slice allocation.

### Motivation

We identified the root cause in the orchestrator check. The collectorBundle is initialized first, and then we check whether the pod is the leader. If it's not the leader, the check is skipped.
https://github.com/DataDog/datadog-agent/blob/main/pkg/collector/corechecks/cluster/orchestrator/orchestrator.go#L215-L243

However, during initialization we start the informers, whose deleteFunc buffers all deleted resources.
https://github.com/DataDog/datadog-agent/blob/main/pkg/collector/corechecks/cluster/orchestrator/collector_bundle.go#L326-L327

Because the cluster agent pod is not the leader, it never flushes the buffered resources, which causes the memory leak issue.

### Describe how you validated your changes

Deploy the cluster agent with multiple replicas, then create a large number of custom resources (e.g., karpenter.sh/v1/NodePool). Delete all of them and repeat this operation several times, while monitoring the memory usage of the cluster agent pods.

### Additional Notes

Co-authored-by: kangyi.li <kangyi.li@datadoghq.com>
(cherry picked from commit cec04dd48a2edbd43cc69b47fde93343113042cd)