AGTMETRICS-385 refactor metrics payload creation and routing (#42852)
### What does this PR do?

Currently, we build payloads with minimal information about their intended destination, and later pair them up with resolvers (for autoscaling failover) or filter them out in the domain forwarder (for MRF).

This PR extends the pipeline mechanism found in the serializer to also keep track of each intended destination for each type of payload we build: local, MRF or regular metrics payload.

When pipelines are prepared to build the payloads, we combine runtime information with configured destinations (in the shape of domain resolvers) and track explicitly each destination (resolver and endpoint) where each particular payload should go. Because MRF may or may not use the same payload as the regular route, pipelines are deduplicated based on the filtering rules.

Since each pipeline in the serializer knows where the payloads should go, it seemed reasonable to build transactions already in the pipeline and hand them off to the forwarder, and the latter should have a generic API to accept complete transactions.

### Motivation

This prepares groundwork for adding more variation to payloads we build based on properties of each domain, which would be clunky to encode as `Destination` property of the `BytesPayload`.

### Describe how you validated your changes

Configure agent with MRF and two fake intakes (one regular, one MRF).

Observe metadata payloads going to both, and metrics payloads only to the primary one.

Call `agent config set multi_region_failover.failover_metrics true`, and observe identical payloads on the MRF intake.

Configure `multi_region_failover.metric_allowlist` and repeat above, making sure that MRF intake has filtered payloads.

### Additional Notes

MRF and Autoscaling failover do not agree on the semantics of empty allow list. If empty, we used to build an `AllRegions` payload, that was handled by MRF destination without additional filtering. Autoscaling failover payloads had the same logic in the serializer, but the forwarder would later drop them as only `LocalOnly` payloads were converted into transactions for the resolver with Local destinationType flag. Specific behavior for each of the two special payload types is now handled by `buildPayloads` directly and methods that create filters reflect different behavior explicitly.


Co-authored-by: vikentiy.fesunov <vikentiy.fesunov@datadoghq.com>