[ AGNTLOG-349 ] streamline the agent status checks during the file_tailing e2e tests (#43160)
### What does this PR do?
The log e2e utilities have been expanded with the capability to reliably detect when an agent's tailers have started. This is quite useful for some tests, such as the file_tailing test suite, which are racing the tailer's pickup of the file they are testing with the writes they are expecting the tailer to pick up. 

The expectation is that this new functionality will reduce the incident rate of test failures currently being experienced in the file_tailing test suite.

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: ryan.hall <ryan.hall@datadoghq.com>