AGTMETRICS-394 Handle bucket reset gracefully (#43135)
### What does this PR do?

When histogram monotonic bucket (as used by openmetrics check) drops below previous value, treat it as a reset, and apply the same logic as for monotonic_count: if FlushFirstValue is true then use the value as is, otherwise record silently and wait for the next value to compute the difference.

### Motivation

### Describe how you validated your changes

Run an openmetrics app that exports histograms and collect metrics from it with `send_histograms_buckets` enabled. After a few metric points are reported, restart the app. Agent should not emit a warning and there should be no gap in metrics in the app.

### Additional Notes


Co-authored-by: vikentiy.fesunov <vikentiy.fesunov@datadoghq.com>