[CONTINT-4861] Filter out events from pause containers when using containerd collector (#43145)
### What does this PR do?
This PR addresses a bug where events were being emitted from user-created `pause` containers when using the `containerd` collector. This is inconsistent with the behavior from the docker events collector.

We do this by passing in a `pauseFilter` to the event subscriber. 

The bug does not apply to K8s-managed pause containers, which are marked as `sandbox` containers by the runtime. 

### Motivation
https://datadoghq.atlassian.net/browse/CONTINT-4861

### Describe how you validated your changes
Built the datadog agent, deployed onto a `minikube` cluster with the containerd runtime. Created a `ubuntu` container and a `pause` container, and saw that only events related to `ubuntu` were emitted.

<img width="1400" height="292" alt="image" src="https://github.com/user-attachments/assets/abc80373-3167-4c30-bb36-483ed27573b0" />

### Additional Notes


Co-authored-by: jon.rosario <jon.rosario@datadoghq.com>