[ AGNTLOG-105 ] Remove extraneous systemd e2e testing (#43001)
### What does this PR do?
Several of the journald system tests were well covered in the tailer_test.go equivalent, and could be safely removed from the e2e suite. Included in these removable tests are the ones responsible for the test suite's flakiness, which happened when journal logs were written before the first journal tailer had been initiated. The one remaining test should not be vulnerable to this error, but a slight change was made to the agent status check to confirm the tailer is present anyway. The hope is that this change will encourage any test expansions in this area to do the same.

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: ryan.hall <ryan.hall@datadoghq.com>