Wrap visual studio installation in a repository rule (#43177)
### What does this PR do?

It implements a repository rule to work with a local Visual Studio installation, and makes the Windows Python build use it.

### Motivation

[ABLD-179](https://datadoghq.atlassian.net/browse/ABLD-179)

This is a follow up to https://github.com/DataDog/datadog-agent/pull/42848 (the Python build for Windows) in an attempt to make Bazel more aware of the VS tools being used for that Python build which runs based on MSBuild.

### Describe how you validated your changes

Locally and on CI.

### Additional Notes

I also explored the idea of having Bazel install Visual Studio and try to fully manage it like that. Even though I got that to work at one point, there were two significant caveats:

- Path length issues: The installation would only work with an output root of `C:\bzl` and not, for example `C:\bazel\bzl`, despite the OS-level setting. This would cause brittleness.
- Visual Studio installations, even if "isolated" to some extent (a feature added in VS 2017), still write configuration and stuff to directories out of Bazel's purview (under `%LOCALAPPDATA%`). I think this could be enough to make it mostly unfeasible to make it work with Bazel caching.

The way this PR implements things, I believe that the Python build is seeing the msbuild.exe executable as an input. This is probably _good enough_ for now, even though it doesn't guarantee that a lot of other things are equal. We can of course revisit all of this as requirements change or are better understood.

This PR's implementation is meant to provide enough flexibility such that it should be possible to build outside of the build image by supplying the appropriate environment variable to point to an existing installation.

[ABLD-179]: https://datadoghq.atlassian.net/browse/ABLD-179?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Co-authored-by: alex.lopez <alex.lopez@datadoghq.com>