Fix ReplaceBindParameter obfuscation config handling in python dbms checks (#43197)
### What does this PR do?
We previously exposed the ReplaceBindParameter config in obfuscate module from go-sqllexer [here](https://github.com/DataDog/datadog-agent/pull/41408) but it was not fully wired up correctly in datadog_agent.go for use within python checks. 

The first commit in this PR fixes the bug by manually mapping the missing config field. This is brittle and fully duplicates the config struct in two places. The second commit embeds the `obfuscate.SQLConfig` struct directly inside the `sqlConfig` wrapper in `datadog_agent.go`. This should help simplify changes needed in the future when exposing new config options from go-sqllexer and ensure it works the same for all usage of the obfuscate package within the agent.

### Motivation
The original commit is needed to fix a bug raised in a [support ticket](https://datadoghq.atlassian.net/browse/SDBM-2179)

The second commit was made in order to cleanup and simplify obfuscation config handling within the datadog_agent checks bridge.

### Describe how you validated your changes
These changes were validated by building a local build of the agent and running it in staging with ReplaceBindParameters=true. Bind parameters are now being obfuscated properly with this patch and existing obfuscation config options are still working as expected

SqlServer query with bind parameters (`@1`) not being obfuscated properly on Agent v7.72.2
<img width="869" height="461" alt="Screenshot 2025-11-18 at 3 47 44 PM" src="https://github.com/user-attachments/assets/652fb80d-f1da-4b17-87a1-cd194dc22491" />

And here's the same query properly obfuscating the bind parameters on the custom image built from this branch
<img width="832" height="467" alt="Screenshot 2025-11-18 at 3 47 18 PM" src="https://github.com/user-attachments/assets/960de806-70ec-4bc7-a3ba-39d567608c39" />


### Additional Notes


Co-authored-by: eric.weaver <eric.weaver@datadoghq.com>