[AGENTCFG-405] Fetching instance-type from agented hosts (#42940)
### What does this PR do?

Refer to [this doc](https://docs.google.com/document/d/1lQiJII2b4R8TzvPElg9Q8ynMqdMg3GHapQ_rVQFmFAQ/edit?usp=sharing). After talking to stakeholders, we have chosen to implement **option 3**; the work for this option on the agent side is to fetch the instance type from various cloud providers and report them in a new field in the inventoryhost payload. This branch contains the agent-side work needed to accomplish this.

### Motivation

### Describe how you validated your changes

We ran an agent on each cloud provider with these changes incorporated, and when we run `agent diagnose show-metadata inventory-host`, we see `instance-type` as a field in the payload.

### Additional Notes

### What does this PR do?

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: s-alad <saad.naji@datadoghq.com>
Co-authored-by: dustmop <me@dustmop.io>
Co-authored-by: rahul.kaukuntla <rahul.kaukuntla@datadoghq.com>