[EBPF-931] gpu: emit high cardinality tags for Docker containers (#43189)
### What does this PR do?

Changes the GPU check so that we collect high cardinality tags for Docker containers. It also refactors the code to take a way the tag retrieval  and caching code to a separate structure for clarity and improved testing.

### Motivation

Docker containers are uniquely identified by `container_id` and `container_name`, without these two tags we cannot link the tags on GPU metrics to specific containers.

### Describe how you validated your changes

Manually validated that the correct tags are emitted, unit tests included too.

### Additional Notes

Co-authored-by: guillermo.julian <guillermo.julian@datadoghq.com>