CONTINT-4886 CONTINT-4899 Add CRD to kubeapiserver collector
Add a new CRD Entity type.

Watch for all CRDs on the cluster.

Add them to WorkloadMeta when found with their latest used version.

This will be useful later for autodisovery to trigger KSM checks based
on the CRD type we find.

In WorkloadMeta it will be available to any other future components that
need it.

co-author: Mathew Estafanous <mathew.estafanous@datadoghq.com>

Signed-off-by: Alexandre Lavigne <alexandre.lavigne@datadoghq.com>