[AMCC-158]dogstatsd workers share filterlist matcher (#43232)
### What does this PR do?

The dogstatsd workers just have a pointer to their matcher so they can share the one instance.

### Motivation

The matcher is never mutated, so there shouldn't be any issues with them all using the same instance. This saves memory when the list is large.

### Describe how you validated your changes

I have tested manually that metrics can still be blocked adding and removing them from the list via remote config. It still works as before.

### Additional Notes


Co-authored-by: stephen.wakely <stephen.wakely@datadoghq.com>