[OTAGENT-580] Add ebpf Config validation (#43214)
### What does this PR do?

Adds validation for the eBPF collector configuration in the host profiler receiver. This ensures that configuration errors are caught early during startup.

### Motivation

### Describe how you validated your changes

Tested with valid and invalid configurations to ensure validation works as expected. Do not add an unit test to not have them break if validation change in eBPF collector configuration (another repository).

### Additional Notes



Co-authored-by: olivier.gaca <olivier.gaca@datadoghq.com>