dyninst/loader: fix bug with circuit breaker stats (#43323)
### What does this PR do?

Before this change we were interpretting the number of nanoseconds as the number of hits. This lead to quite incorrect logic with regards to the circuit breaker firing (it would fire too often).

### Motivation

I noticed the circuit breaker tripping when it absolutely should not have been.

### Describe how you validated your changes

Added a test.

### Additional Notes

Relates to https://datadoghq.atlassian.net/browse/DEBUG-4202.

Co-authored-by: andrew.werner <andrew.werner@datadoghq.com>