errors: Support wrapped errors (#43285)
### What does this PR do?

Changes the `reasonForError` logic to support wrapped errors.

### Motivation

If an error coming from the constructors in the package is wrapped, the rest of the `IsX` functions will not work. 

### Describe how you validated your changes

Unit tests added to cover the new case. I checked existing usage of the `IsX` functions to ensure that we are not activating code paths that weren't being used before. `IsNotFound` is the only one used more broadly, with the rest being either unused or only in tests.

### Additional Notes

I found some places in the code that are using this agent-specific logic to check for errors that come from other packages:
* `pkg/clusteragent/appsec/istio/detection.go:21` 
* `pkg/util/kubernetes/apiserver/controllers/hpa_controller.go:157` 
* `pkg/util/kubernetes/apiserver/controllers/wpa_controller.go:209` - although there's a check above it for `err != nil` already

Co-authored-by: guillermo.julian <guillermo.julian@datadoghq.com>