Store attr in S3 bucket (#43339)
### What does this PR do?
Adds attr package to our S3 bucket to ensure resilience in case original
source is unavailable. The order in `urls` attribute in `http_archive`
is important, so S3 URL should always come first (if we assume our
bucket is more reliable).

### Motivation
The original source of attr is currently unavailable and it breaks our
builds.