Shrink context map when possible
This commit is a part of response to #incident-45849. I'm tinkering to
reduce -idle memory consumption.

I've added a 'shrink' to contextResolver which cuts the size of the
interior map in half when contexts expire if the map contents is
smaller than half its capacity. This will penalize setups where
context counts vary rapidly and benefit setups where context counts
are roughly stable over time. A similar tactic is taken for tags
store.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>