usm: native-tls: Fix a leak
Everytime SetFD method was being called we created an entry in ssl_sock_by_ctx, but if we failed to
associate a connection with the SSL context, we would have skipped the cleanup.

The scenario happens mainly in nginx, and resulted in leaking entries in the map.
The fix ensures we clean the map even if we don't have a connection associated with the context